Added a lot of stuff:
A "Pn532" class has been added to keep all of the NFC stuff in one place.
This class is still very minimalistic and is neither bug free, at the moment of writing this I cannot get an ACK back from the PN532 ("0x00 0x00 0xff 0x00 0xff" I believe).

Also added some images, these will be shown on the promotional part of the screen.

At this stage of the project the code is pretty lackluster in terms of comments, these will be added later on down the line.